Crate get_if_addrs

get_if_addrs

Structs

Ifv4Addr
Details about the ipv4 address of an interface on this host

Ifv6Addr
Details about the ipv6 address of an interface on this host

Interface
Details about an interface on this host

Enums

IfAddr
Details about the address of an interface on this host

Functions

get_if_addrs
Get a list of all the network interfaces on this machine along with their IP info.