Crate get_if_addrs

get_if_addrs

Retrieve interface IP addresses in windows and posix systems

Structs

Ifv4Addr
Details about the IPv4 address of an interface on this host

Ifv6Addr
Details about the IPv6 address of an interface on this host

Interface
Details about an interface on this host

Enums

IfAddr
Details about the address of an interface on this host

Functions

get_if_addrs
For non-Windows operating system, use this function to get address